Various mostly minor fixes (leaks/crashes)
Memory leaks, FILE leaks, uncheckd FILE *, unchecked allocations,
range checking, free()ing stack pointers, etc.

Make file/compressed sizes unsigned and use UINT32_MAX for LFN size.
Avoids casting negative values to size_t.

Deal with sizeof(long) == 4 on Win32

Allow seeking using SEEK_CUR with a uint32_t when LONG_MAX < UINT32_MAX

+++ src/doors/clans-src/src/myibbs.c
@@ -89,7 +89,7 @@ static void GetMessageFilename(char *pszMessageDir, uint32_t lwMessageNum,
 {
 	char szFileName[FILENAME_CHARS + 1];
 
-	snprintf(szFileName, sizeof(szFileName), "%" PRId32 ".msg", (uint32_t)lwMessageNum);
+	snprintf(szFileName, sizeof(szFileName), "%" PRIu32 ".msg", lwMessageNum);
 	MakeFilename(pszMessageDir, szFileName, pszOut, sz);
 }
